code corrections

+++ Database/dbo/Tables/University.sql
@@ -1,10 +1,12 @@
 ﻿CREATE TABLE [dbo].[University] (
-    [Id]          BIGINT         IDENTITY (0, 1) NOT NULL,
+    [Id]          INT            IDENTITY (0, 1) NOT NULL,
     [Name]        NVARCHAR (50)  NOT NULL,
     [Description] NVARCHAR (100) NULL,
     PRIMARY KEY CLUSTERED ([Id] ASC),
     UNIQUE NONCLUSTERED ([Name] ASC)
 );
+
+
